requirements.txt
asn1crypto==1.5.1
asn1tools==0.164.0
autopep8==2.0.2
bcrypt==3.2.0
bitstruct==8.15.1
capstone==5.0.0rc2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colored-traceback==0.3.0
cryptography==36.0.2
Cython==0.29.30
diskcache==5.4.0
factordb-python==1.3.0
gmpy2==2.1.2
idna==3.3
intervaltree==3.1.0
libnum==1.7.1
Mako==1.2.0
MarkupSafe==2.1.1
mpmath==1.2.1
numpy==1.22.3
oscrypto==1.3.0
owiener==1.0.6
packaging==21.3
paramiko==2.10.3
Pillow==9.1.0
plumbum==1.7.2
prompt-toolkit==3.0.33
psutil==5.9.0
pwntools==4.7.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.14.1
pyelftools==0.28
Pygments==2.11.2
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.8
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
pywin32==303
requests==2.27.1
ROPGadget==6.7
rpyc==5.1.0
six==1.16.0
sortedcontainers==2.4.0
sympy==1.10.1
tomli==2.0.1
unicorn==2.0.0rc6
urllib3==1.26.9
wcwidth==0.2.5
